move ScannerImpl util classes into same file
